+++ src/main/java/com/monds/nextstep/lotto/domain/LottoMachine.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package com.monds.nextstep.lotto.domain;
 
 import static com.google.common.base.Preconditions.checkArgument;
